+++ RELEASENOTES.rst
@@ -4,6 +4,7 @@ Release notes
 1.0.30
 ------
 
+- Title parameter added to PrintProgress
 - Repeat now repeatably calls functions
 - FilterCol added
 - CSVReader now supports format functions per column